QE: Cucumber step definitions shouldn't contain other than step definitions
