+++ testsuite/features/step_definitions/api_common.rb
@@ -6,23 +6,6 @@
 require 'json'
 require 'socket'
 
-def reset_api_client
-  ## Testing inside containers needs to be done without ssl
-  ssl_verify = if $is_container_provider
-                 false
-               else
-                 true
-               end
-
-  $api_test = if $debug_mode
-                ApiTestXmlrpc.new(get_target('server').full_hostname)
-              else
-                product == 'Uyuni' ? ApiTestHttp.new(get_target('server').full_hostname, ssl_verify) : ApiTestXmlrpc.new(get_target('server').full_hostname)
-              end
-end
-
-reset_api_client
-
 ## system namespace
 
 Given(/^I want to operate on this "([^"]*)"$/) do |host|
@@ -406,14 +389,6 @@ When(/^I wait until there are no more action chains$/) do
 end
 
 ## schedule API
-
-def wait_action_complete(actionid, timeout: DEFAULT_TIMEOUT)
-  repeat_until_timeout(timeout: timeout, message: 'Action was not found among completed actions') do
-    list = $api_test.schedule.list_completed_actions
-    break if list.any? { |a| a['id'] == actionid }
-    sleep 2
-  end
-end
 
 Then(/^I should see scheduled action, called "(.*?)"$/) do |label|
   assert_includes(
